Fix bua when using additional variables (e.g. wtots1)
This is not affecting reconstruction, since wtots1 is not used presently. If the
MVA was trained with additional variables than the usual and the input weights
are done with an old version the tool, the tool need to manually parse some strings to
know which variables are needed. To do this parsing the tool needs to know which
are the variables that may be used: wtots1 was missing.


Former-commit-id: b9e186f0d2c6b0c7f8d426cc83c2b42b0b2d53c3