Update to the sTGC Geometry
Chilufya's fix to the sTGC geometry to include the honeycomb layers in the quads.  This fix should cause the Z positions of the simulated hits to line up correctly with the real planned detector positions as well as provide the correct Z positions for the wire planes used in the Digitization.


Former-commit-id: 5db1537d1cf16c1d1e4345ece22f074eb8727dcf