Switch to using `InputConverter` to convert `GenEvents` into  `G4Events`
These changes move from using `FADS::GeneratorCenter` inside the Run Manager
objects to read in the `McEventCollection` and also to create the output
`McEventCollection` to using an `SG::ReadHandle` and an `SG::WriteHandle`
inside `G4AtlasAlg` to make the data dependencies explicit.

Within `G4AtlasAlg` the `InputConverter` service is used to produce the `G4Event`
rather than the `FADS::GeneratorCenter` code.

Initial tests show that the simulation is unaffected by these changes, but there
are some changes in the massses recorded for truth particles in the output
`McEventCollection`. This still needs to be understood. See ATLASSIM-3110 for
more details.


Former-commit-id: fa834b16358c5c9479d6fd762ae6672cd32b74f1