Modification of RunTier0Tests.py in preparation for use in automatic CI
A few minor edits are made in order to have the output be slightly more
meaningful in the cases where the script is being run as part of the CI tests
(the script will now print all branches that are changed, instead of prompting
the user to look themselves).  There were also so minor bug fixes (reporting
a memory reduction as a failure, instead of a warning), a summary of
whether all tests passed or failed, and finally making the logging
messages reflect the warning/error/info state.


Former-commit-id: f1236812f103a86d8964fb1288d1015326e29395