Add StoreName to DataHandle keys
Need to distinguish between DataObjIDs that are associated with different
stores, as two objects with the same key can live in different stores, and
without the store name, the Scheduler thinks these objects are the same.

DataObjIDs will now look like "storeName:StoreGate_key". VarHandleKey::key() returns
just the StoreGate_key. This is held as a data mamber of the VarHandleKey class,
and get updated appropriately by the VarHandleKey::parseKey class when the
handle is constructed or assigned.

Compound keys with "/" can only be put in the ConditionStore.

Blank keys are allowed, and the DataObjIDs are also blank (ie, no storeName).

StoreID.h now contains a repository of default store names, and the association
between the name and the StoreID. This allows the removal of hardcoded store
names like "StoreGateSvc".


Former-commit-id: 3f9e9addeddf3518c65cf1652cf89835994b88e5