Fix issue with the file closing and unreadable histograms if deleting
If the option to write a new file with the unmatched histograms
deleted was used, the resulting file was unreadable.

This fix reads the lumiblock information if available, but if doing
this, the file takes a *very* long time ( ~ 2 minutes ) to close.
This removed it from the root list of managed objects so that it
closes in a reasonable time.


Former-commit-id: bd73f4095dd50f3c6eb1d96a05c2a9c662be6629