JetRec: Avoid corruption of derived jet containers due to thinning.
Changes discussed in ATLASRECTS-3820 introduced thinning
for CaloCalTopoClusters.  The containers LCOriginTopoClusters
and EMOriginTopoClusters derive from CaloCalTopoClusters via 
a shallow copy, so if CaloCalTopoClusters is thinned, then these
two containers also need to be thinned in the same way, or they'll
be corrupt in the output file.

Schedule thinning for these containers if needed.

Not sure if this is really the best place for this, but it seems to work.

This should possibly go to r21 as well.



Former-commit-id: 2a948d3807e07baeaa11d51f6ad0595115b5c8f1