Reject standalone muons in the case that beamline extrapolation fails
This was first noted in the context of ATLASRECTS-3997, where standalone muons were created without track particles because the beamline extrapolation failed.
This can happen in rare cases because the tracks that form the standalone muon are created by extrapolation to the beamspot instead.
Goetz addressed the immediate problem with a change to the CaloIsolationTool.
However, there is no reason to accept these muons, as the ME track particle for the standalone muon is also the primary track particle.
So this fix checks for cases in which the track particle is not created and rejects those muons.


Former-commit-id: 9ad24df501c4ea0865ce693188a31e069b243ea6