TSOS sorting modification
As seen in ATLASRECTS-4005 (and other JIRA tickets), in rare cases strange kinks or other unusual features in tracks can cause TSOS's to be added out of order.
This happens in the hit recovery stage: hit recovery creates a list of hits that may possibly be out of order, so it has to sort them.
Usually this is a straightforward process, but for low-pT tracks, or tracks with kinks, there is the potential for problems.
These are quite rare, but as they often lead to crashes they are fairly serious.
In almost all cases, the issue is that changes in track direction in the calo lead to MS and calo hits being interchanged.
The MS and ID hits are almost never added out of order with respect to each other, as the track direction rarely changes much in the MS or ID.
So the solution is, when sorting the hits, to always look first at which volume the hit is in.
If the two hits are in the same volume, they are sorted by distance as usual.
If they are in different volumes, then we use the volume to sort them: previously, this was only done if the distance sorting broke down.
I tested this successfully on the problem examples in ATLASRECTS-4005 and ATLASRECTS-3974, so it should reduce warnings as well as avoid crashes.


Former-commit-id: 32a0378922057669c3ea276672b82259222d72d1