Reworked setting up datahandles in JetMomementTools.
The strings passed to decalreProperty in the Tool constructors were used as storegate keys.
An attempt was made to maintain th string attribute of the tool, and use the sam python attribute
to initialise the new data handle keys. This is wrong - a pythonattribute cannot initialise more than
one C++ attribute. Now use only ot initialise the data handle keys. The string can be retrieved (eg for error massages
using DataHandleKey::key(). The string attributes have been removed from the heasder files.

The ReadHandleKeyArray attribute of JetConstitFourMomTool is initialised to be empty, and filled
in JetConstitFourMomTool::intialise from the vector of strings passed to via the call to declarePoperty
ie from m_altColls;


Former-commit-id: 96ce5b6a4c3b8913ae70b822992a802c172a059b