Fix for ATR-16605
The standalone muon track fit uses the GlobalChi2Fitter (unlike the combined track fit which uses the iPatFitter).
In rare cases (and only when the GlobalChi2Fitter is used for muons) the call to obtain the list of downstream scatterers from the extrapolator gives bad results.
The extrapolator returns a list of thousands of scatterers, which is clearly wrong.
This results in the fitter trying to solve a thousands by thousands matrix, which then leads to trigger timeouts.
To avoid this behavior, a configurable setting is added to the GlobalChi2Fitter to check for the return of over 1000 scatterers and simply quit if so.
It is false by default, but set to true in the muon instance of GlobalChi2Fitter, configured in MuonRecTools.py.
The underlying cause of this problem should probably be studied by an extrapolator expert, but for now this will get rid of the timeouts.
Results of the trig tests are available at /afs/cern.ch/work/w/wleight/public/MuonSW/ATR-16605/.


Former-commit-id: 14e2a4b6920d97a577c28f128d7fac343720ff74