SGInputLoader: allow updating by Scheduler
SGInputLoader can receive its input lists either explicitly via its
Property "Load", or implicitly when the Scheduler marks it as the
DataLoaderAlg. In the latter case, all unmet input data dependencies
are attributed to the SGInputLoader. The union of these two sets
are loaded from StoreGate during execute.

If the loading of a proxy fails during execute, the SGInputLoader
will return a StatusCode::FAILURE, unless the propery FailIfNoProxy
is set False (default is True).

We can also control whether the data obj deps attributed to the
SGInputLoader by the Scheduler are actually loaded by means of the
Property "LoadExtraProxies" (default is True).


Former-commit-id: 16ec8bcfd694d75c3e1b4d404772aa124dc3e56c