Zip or tar multiple small outputs with Archive_tf.py (ATLASJT-296)
The functionality of Archive_tf.py has been expanded, not only
to archive files into a tar format (with or without compression),
but also to zip files. This functionality was requested by ADC to
manage large numbers of small outputs, which need to be stored for
metadata purposes. The extended functionality allows to zip together
the small output files before storage.


Former-commit-id: 24d16b6ef51f5a04ac23c3adebaeac5ff5f08dbe