Athena: Update externals tag to pick up recent static checker updates.
Updating commit-id for externals to pick up some recent static checker changes.
The commit used has only the checker changes when compared with the
current commit; other changes in externals are not picked up by this change.

The checker changes are also in the master branch of externals, so it should
be fine to update to the tip of master when we want to pick up the
other changes from externals.



Former-commit-id: c2276655e8990cf8d7414850420f149d1a3c8040