Further tidying and robustness improvements for InputConverter
Used unique_ptr in cases where objects created on the stack could
possibly be leaked. Made the creation of G4PrimaryParticles much
stricter. If there is a problem then nothing will be added.

Tried to further improve the issue with small changes in GenParticle
mass after switching to this method of G4Event creation. I'm pretty
convinced that this is just due to rounding errors now and that
there's no actual bug in the code.


Former-commit-id: 7050d7be35b4eac16a7024e29e40f1ae8308dedb