Sweeps from SVN and modifications to support production of Higgs DAODs
ATLASG-1202

This is to enable the HIGG DAODs to be run in 21.2 AthDerivation. It brings in the following tags from SVN:
* TruthConverters-00-00-04
* DerivationFrameworkHiggs-00-05-32

The latter is further modified to make it compatible with modifications to `IHiggsTruthCategoryTool` in TruthRivetTools (see ATLASG-1247, !2985). Additionally DerivationFrameworkMuons is modified, adding the `MuonTriggerList.py` file.

With these changes, all of the HIGG DAODs should be able to run on 21.0 data.


Former-commit-id: 9dbc75c092e29387a0704ecf180f92704d76c511