Updated to xAODBPhys-00-00-15; moved tools to PhysicsAnalysis/BPhys/BPhysTools
This update to xAODBPhys adds the light-weight BPhysMetaDataHelper class
which helps with reading B-physics specific metadata information atteched
to FileMetaData objects written by certain BPHYx derivations.
A new package PhysicsAnalysis/BPhys/BPhysTools is added to contain
dual-use tools for BPhysics.

Please note that the tool class BPhysTrackVertexMapTool has been moved
from the xAODBPhys package to the new PhysicsAnalysis/BPhys/BPhysTools
package as recommended by Attila K..  This is a modification w.r.t.
to the contents of xAODBPhys-00-00-15 in SVN.

Important: When the trunk/latest tag of DerivationFrameworkBPhys will be
imported from SVN to git, a dependency on the
PhysicsAnalysis/BPhys/BPhysTools package needsto be added and the
include statements for the BPhysTrackVertexMapTool class will need to
be adjusted in BTrackVertexMapLogger.h.

Former-commit-id: b07ceb3115bd9708d196ad3f45baf7ed592d6894