Made asg::ToolStore keep thread-specific lists of tools.
This is a first attempt at making analysis tools thread-safe. In this
implementation every analysis tool needs to have a separate instance
in every thread. The asg::ToolStore code will have to be made a little
smarter later on to allow global (thread-safe) analysis tools to be
used as well.


Former-commit-id: c8f55f931ccb95152d55ac7ecc798fa68f5351d6