AthenaBaseComps: Fix problems with var handle key arrays.
A VarHandleKeyArray is a vector of key objects.  Each of the contained key objects
needs to be declared to the algorithm/tool for dependency analysis to work.
This was being done from a Gaudi property update handler.  This however has
some problems:

 - If the member variable is modified directly from the class, then the update
   handler won't get called.
 - More importantly, the declaration results in pointers to the key instances
   being saved in the algorithm.  If the vector is later enlarged, then these
   pointers may become invalid.

To fix this, we defer declaring the key objects contained in arrays until after
initialize() is complete.  For dependency analysis to work correctly, we also
need to extend inputHandles() and outputHandles() to explicitly include
the array keys.


Former-commit-id: 630a427109873e230a40a6138ee35faa42e3e156