Tream L1_MU20,L1_MU21-based chains in a special way (ATR-16612)
Added new methods to ChainGroup to handle prescale calculation of chains with multiple
L1 items for which the correlation of firings is known. This is currently only the case for
chains that are seeded by L1_MU20,L1_MU21.

In the future, if new chains are added with that property, the logic on how the prescale is
calculated must be added to the correlatedL1Prescale method.


Former-commit-id: 5504d6473bf32025c0e4d69b28265c07bd8e8c70