TrigMuonHypo: const fixes.
Due to a long-standing DataVector bug, it has been possible to get a
non-const pointer from a const_iterator.  Numerous packages have relied
on this behavior, making it difficult to fix.

This package relies on it due to calling the non-const legacy method
TrigMuonEFInfo::CombinedTrack().  Rewrite these calls so that they
use only const methods.



Former-commit-id: 55a13e2f7b64574033809443ba190e52c7c3cb42