add EL::Algorithm::evtStore() method
This method should reasonably well mimick the
AthAlgorithm::evtStore().  The primary motivation is to allow using
the same instructions for EventLoop and Athena in the tutorial and
other documentation.  In particular the various instructions for CP
tools and using xAOD objects often only differ in how the input
objects are retrieved.

Former-commit-id: a10b7ce1c8e5140a50eafc8ba3ba37a2c2acc290