Updated AthDerivation/version.txt to 21.2.0.0.
This is to differentiate the Gaudi version built by AthAnalysis and AthDerivation. With the idea being that "normal" AthDerivation releases will have versions 21.2.1.0, 21.2.2.0, etc. after this. The fourth digit would be used only if an older release needs to be patched for some strange reason. See ATLINFR-1830 for further details about this.

Former-commit-id: ac04da726e95261f89f42f1ba3445a970df653d4