Adding empty PRD containers if we don't decode anything.
It can happen in the trigger that the RoI doesn't overlap with a given detector technology, so we
don't decode anything, and so don't ever record a PRD container. Further downstream it is assumed that
these containers exist for all four muon technologies. For the moment, I've simply made sure that we
record an empty container if we don't decode any part of the relevent detector. This will be completely
transparent for offline reco.


Former-commit-id: da9feec673433d2260641c6cfb09b9297b2f0304