Remove all usage of MCTruthMenu and MCTruthStrg classes in python
There are a number of job options which define callbacks to tweak the old-style
MCTruth Strategy configuration. This code is all obsolete. For now I've left it
commented out. The next stage will be to check for (and create if needed) the
equivalent ISF truthStrategy configurations and mention the error messages
that these should be used. Force jobs to fail if obsolete configuration is used.


Former-commit-id: 990bd90feef663f322318b799210e3cfa77f8024