Cleanup unused classes from G4AtlasAlg package
The `AthenaHepMCInterface` and `TruthHepMCEventConverter` classes
have been replaced by the `InputConverter` service from `ISF_Services`,
so it is safe to remove these two. This also removes all dependencies
on the `SimHelpers` and `FadsKinematics` packages, so these two can be
dropped from the `CMakeLists.txt` file. These changes are necessary
for ATLASSIM-2771, which aims to remove the `FadsKinematics` package
from the release.


Former-commit-id: 6584252b8fa6cb6c53d27e86d2a33ac65f404653