Cleanup unused classes from G4AtlasAlg package
The `AthenaHepMCInterface` and `TruthHepMCEventConverter` classes
have been replaced by the `InputConverter` service from `ISF_Services`,
so it is safe to remove these two. This also removes all dependencies
on the `SimHelpers` and `FadsKinematics` packages, so these two can be
dropped from the `CMakeLists.txt` file. These changes are necessary
for ATLASSIM-2771, which aims to remove the `FadsKinematics` package
from the release.


Former-commit-id: e2ebe8749e96ba72ac61d018c8ee1ccf27d13fc4