SGInputLoader: do not require storename in explicit object list
Will automatically insert the default EventStore name into the container
of objects to auto load if it's not already there.

Will check if store is default EventStore, and prints a warning message
if it's not. This warning message will only appear ONCE, at the first event,
when the full list of proxies to load is assembled. Will not attempt to load
proxies for non-EventStore data.

Also uses initializer list syntax for Properties.


Former-commit-id: 52af9148726d5707b232ebf7a48c703623b4fd27