Introduced copy2DFEval for PixelMon2DMapsLW class
Merging the two classes into one makes it necessary to have a switch for
controlling the FE-copy method that is used for IBL for certain
occasions. Since the non-LW class did not have this copy method
implemented, but LW did (without a switch), we now need this switch to
turn the functionality on and off. For all existing 2D maps containers
this switch was set to "true".

For the new objects, the classes of which were changed to LW, this
switch has to be "false", otherwise the bin entries are different
compared to before the changes.


Former-commit-id: 0fc6ada3f7034efb33ec059ebd29ff0a43c0a801