Use ReadHandle in InDetPRD_AssociationToolGangedPixels (ATLASRECTS-4026)
In InDetPRD_AssociationToolGangedPixels from package
InnerDetector/InDetRecTools/InDetAssociationTools, replace calls
to evtStore()->retrieve() by ReadHandles.

In this class, a failure to retrieve the collection was not treated as
fatal, with a message stating that the failure was normal if the pixel
is off. In this case the container pointer remained unititialized and
was presumably not accessed in the member functions because there are
no PRD. This commit just replaces the member pointer by a ReadHandle,
which gets assigned only if the ReadHandleKey initialization
succeeded. Now, if for any reason the collection is accessed after a
failed initialization, the code will throw an exception rather than
segfaulting.

There is one changed bit of logic: before this patch, if the container
was expected to be in the store but the retrieve() call failed, the
code showed an error message but didn't fail. Now, if this happens the
call to makeHandle will throw an exception.


Former-commit-id: 2cc2f2b99f7e7769fdf6701a7db4a168557bf6f6