Tidy ISFG4GeoHelper class
The main change is to switch the `checkVolumeDepth` method back to throwing `G4Exceptions`
rather than string-literals. Some commented out code is left in the class as the next step
is a re-write of part of the main client of this method `TrackProcessorUserActionFullG4` to
determine the `truthVolLevel` at runtime rather than having to pass it in as a configurable
property. Code to do something very similar already exists in the `RecordingEnvelope` class.


Former-commit-id: 70b3f65bda1be256251375aea01a44fd026a3be3