fix of decoding for BMG chambers
The initial decoding of the BMG chambers was requested in
MR !1726 . The data quality group (@alsydore , @sawebb )
spotted that only BMG2 histograms are filled. Initially it
was thought the issue was caused by naming complications
coming along with these chambers. However, it turned out
that there was a bug introduced with the new decoding.

The BMG chambers are read out by HPTDC chips. A feature of
these chips is that they do not send a trailer word while
this was sent by the common AMT chips used for MDTs. The
decoder uses the trailer word as condition to exit the hit
decoding loop. In consequence the decoder like it is running
currently is decoding a full ROD (which has a BMG) and put
all hits into the first BMG chamber only. The affected RODs
hold the 12 BMGs and the 4 elevator chambers. So all the
data of these chambers is just put into the BMG2 chambers.

The code requested with this MR will in case of the BMGs
also interpret the dataword in terms of a CSM dataword.
As after a trailer word from the chip a trailer word of
the CSM follows the CSM trailer is used to stop the
decoding loop for hits in order to jump to the next
chamber. A local test confirmed that now all histograms
are filled and that only the decoding of the affected
RODs shows changes. The change was also discussed with
@rosati and @pfleisch who support the MR.

This MR would be important to be run at P1 and offline
as soon as possible because it's important for calibration
( @diehl ) and the alignment ( @giraudpf ).


Former-commit-id: d9b683bcd91bb27bbeb69540268a964d9ae80d76