VP1SimHitSystems: Fix compilation error in debug build.
A missing #include was causing this package to fail in the debug build
(and only the debug build due to a NDEBUG #ifdef).

From the comments in SbBasic.h, it looks like it is not meant to be
necessary to #include this file, so this may represent an Inventor bug.
Probably not really worth tracking down at this point.



Former-commit-id: 8a0cdb96ffbee0d09e683fc660de5192fdb42d3d