Fix athenaMP outputs when ATHENA_PROC_NUMBER=1 (ATLASJT-365)
This brach fixes the output of athenaMP when this is activated with
only one processor. In this particular case, an output file with the
correct name was not produced. The single-core worker output was
symlinked correctly in the top directory, but it was not self merged
(or moved) and renamed to the output file name provided by the
user. In this branch, it was chosen to move the single worker output
file to the top directory and rename it with the output file name
given as an argument to the transform.


Former-commit-id: 4a427bfb3063401c7ee219009ae9844302d2fb44