AthenaKernel: Conditions should be indexed by run+lbn, not run+event.
Run-based conditions should be indexed by run+lbn, but the condition handle
code was using run+event instead.  Fix so that run+lbn is used.
To really do this right, EventIDBase needs to be changed in Gaudi
to support ordering via run+lbn.  The comparison wrappers added here
are transitional until the needed changes can be added to Gaudi.



Former-commit-id: 2557a1f7ab85d97807cb6b6f90eb84c180f336b8