update the REALISTIC IBL clustering
updated the centroid calculation because
the edge pixel sensor size of planar and 3D are different.
You can use REALISTIC IBL 3D clustering
by changing the Boolean "IBL3D_REALISTIC".
It affects to the Bank gen and Sector/Const gen,
so need to check the impact of this change.
If there are no problems, we need to fix the Boolean "true".


Former-commit-id: f5f297953df5531251952f7eb2018f412cae622c