The DetectorGeometrySvc configuration should be tweaked as a postInclude
Now that `IRegionCreator` tools are private ToolHandles owned by
`DetectorGeometrySvc` then extra tools should be added after the main
configuration. For this reason I've renamed the related preInclude files
to be postIncludes. `preInclude.Cavern.py` and `preInclude.ScoringVolumeKiller.py`
had to be split into separate pre and post includes. Part of the changes
for ATLASSIM-3379.
It would be nice to find a way to decide whether or not to add
`MuonPhysicsRegionTool` to `DetectorGeometrySvc.RegionCreators` within
`G4AtlasServicesConfig.py`.


Former-commit-id: 667ad53e9cd87b13c19026eecbbac6ad9a046132