Updating to 1 ns merging time
Pursuant to some discussions had at the June 2017 MUON week, this commit changes the merging time resolution of digits to 1 ns which reflects approximately the resolution of the VMM peak discriminator.

It changes the granularity of how the strip VMM array loops over the time series to be in line with this 1 ns.  This may be far too small and increase the run time unreasonably.  This will need to be studied more.

It updates the readTime configuration parameters to reflect the 6.25 ns clock.  This is perhaps not exactly reflective of the physical VMM operation and more discussion with the chip designer as to how the NeighborOn function really works is needed.


Former-commit-id: a6d56f172966411358813b30eaa694f2690da2fb