Updating to get 7/8 tests running
Updating the interface of treatAsYear to allow a run number argument (in
case the user wants to pass a run number instead of having the
GetRunNumber function called).  Adding 2017 as an allowed year.
Updating to only call the muon tools if we are using their year -- this
was causing a crash in one of our make tests.  Updating the pileup
reweighting file to use the latest from MC15c, which gets another two
tests working (the sample for release 21 was not in the old default
file).  Shouldn't change output for users, but should fix tests.


Former-commit-id: 9e358a2f42eb3dfc47f0436ccc62685c518751a4