Allow MuonRawDataProvider algorithms to switch off.
As described in ATLASRECTS-4112, we would like to be able to switch be able to disable
all algorithms from RecExCommon, but still be able to set up needed services. The RawDataProvider algs
are currently always being run, but we can easily acheive our goal if checking also that makePRDs
is true before adding these algorithms (makePRDs is true if running muons normally so this does not affect
the default behaviour).


Former-commit-id: 301f910dec0465474a4d0f63af409cbcdc888e35