Adding a new option for excluding prompt photons from Higgs decays
Option was requested in a discussion of truth derivations.  This
shouldn't affect the vast majority of samples, so I think it's a safe
thing to include in the 'standard' WZ truth jets, which is the reason
for the change to the python file.  The C++ changes are just to
implement the option and to enable it only for photons from Higgs and
MSSM Higgs decays.  For the time being we do not want to try to treat
photons in other types of events (e.g. Vgamma) until we have a clear
prescription for how they ought to be treated.  The photons will appear
in a truth photon collection and can always be treated as jets if the
analysis so desires -- this will just solve a problem for overlap
removal in Higgs analyses with a yy decay.


Former-commit-id: e8428215f5a6e5780802ecca2499c208294d388e