xAODPFlowAthenaPool: Do not call toPersistent from container pool converter.
toPersistent has not been needed for some time, since the same function
happens when the aux store is copied in its pool converter.  Further,
it can interact badly with thinning: if called while the container is in
the thinned state, the aux store can be corrupted.  So just remove
the call to toPersistent.

See ATLJETMET-779.


Former-commit-id: d6142d9244a406b88e2c61d39465192229cf8936