Migrate to the new HepMcParticleLink class
This commit collects the changes in `GeneratorObjects-01-04-00-19`
and `GeneratorObjectsTPCnv-00-06-10-05`. Plus a number of additional
changes:
 - added a method to directly set the `ExtendedBarCode` member
variable of `HepMcParticleLink`, as this was needed by the
`TRTUncompressedHitCollection` TP Converters.
 - added some missing braces to make contents of for loop clearer
 - added the possibility to write old `HepMcParticleLink` format
using event position instead of event number (cherry picked from
commit 31207b97f0453496aaa8794883505c562dee36d4)
 - added a unit test for `HepMcParticleLinkCnv_p2` and extended the
unit test for `HepMcParticleLinkCnv_p1`
 - Changing the `ExtendedBarCode` should force the `HepMcParticleLink`
to re-find the `HepMC::GenParticle`. Previously calling
`HepMcParticleLink::setExtendedBarcode` did not force the `HepMcParticleLink`
instance to refind the `HepMC::GenParticle` to which it should be linking.
This could cause the pointer to be inconsistent with the information cached
in the `HepMcParticleLink`.
 - Ensure that all member variables are properly initialized in the
`HepMcParticleLink` default constructor. Previously `m_particle`
and `m_have_particle` were not initialized properly.

Former-commit-id: 8607d473dddbd5d9921b41f64be2ed52c72a4cdb