Fix for ATLASRECTS-4110
The crash in ATLASRECTS-4110 is due to the fact that the needed set of trackParameters are always deleted by calls to extrapolateToNextMaterialLayer.
Therefore, if nothing works and the code ends up trying the direct extrapolation, the parameters it needs have already been deleted.
In order to avoid this, the parameters are cloned before the call to extrapolateInsideVolume, and the cloned parameters used if necessary.


Former-commit-id: 794bbbed9c0196c264e232cd6f888c8b60e8d9ee