Make the method  TrigHLTJetHypoTool.decide const
This is in anticipation of the tools being called from a reentrant Algorithm.
For this monitoring has been removed (will move to a monitoring service thereby
eliminating the need for to set attributes with the values to be monitored (Run 2 style).
Timing code has been removed (accumulators were used). The JetGrouper is now constructed
every event. This entailed a change to the TrigHLTJetHelper interface. To maintain backward compatibility
thid class has been ropied to TrigHLTJetHelper2. The new Tool code modifies this version.


Former-commit-id: 2dd407cdf2b0f5f6da2125d2b3e2fe25322a971a