Added extra options to the Slurm driver of EventLoop.
The current implementation does not allow enough flexibility
in the configuration to run in some configurations of the batch systems.
This patch adds extra flexibility, in particular allows to add extra commands
to be executed on the worker-node before the main executable is run as well as
the ability to prepend a command to the execution of the main analysis routine,
in order to allow encapsulation in, for instance, shifter or other custom-environment
setup programs.


Former-commit-id: 16884b521fe4166cc494e97202c72ee77205b726