Move non-const member variables of TileGeoG4SDCalc into a struct
The goal of these changes is to make all the methods of `TileGeoG4SDCalc`
const, so that it can be converted into an `AthService`. The first step
is to identify the member variables of this class which are altered many
times during the event loop. These variables are then factored out into a
struct member variable (which will eventually be removed from the class).


Former-commit-id: 06471a4537e6dc6c147fd86a90d5346d3c7d036b