Move non-const member variables of TileGeoG4SDCalc into a struct
The goal of these changes is to make all the methods of `TileGeoG4SDCalc`
const, so that it can be converted into an `AthService`. The first step
is to identify the member variables of this class which are altered many
times during the event loop. These variables are then factored out into a
struct member variable (which will eventually be removed from the class).


Former-commit-id: d8b13b1c775ec18f138a0c563439282c596a80ca