Make TileGeoG4SDCalc::MakePmtEdepTime method const
By making this method take a reference to a `TileHitData` object
as an argument, it no longer needs to explicitly update member variables
of the class (although it is still doing this in practice). One wrinkle
is the the `m_deltaT` variable is still updated and so had to be made
mutable temporarily.


Former-commit-id: 7b803a8aa997b780dc5637e38bbaba69b47641b0