Remove TileHitData member variable from TileGeoG4SDCalc
Next stage in making the `TileGeoG4SDCalc` class const. Now
that the `m_hitData` variable is not explicitly used anywhere
in `TileGeoG4SDCalc` it can be removed from the class and an
independent instance can be created in `TileGeoG4SD::ProcesssHits`
and `TileGeoG4CalibSD::ProcessHits` and passed to the various
`TileGeoG4SDCalc` methods these methods call. In addition, the
`TileGeoG4CalibSD::ScintIDCalculator` method has been updated
to take a `TileHitData` object as an argument, as it too looked
at information in `TileGeoG4SDCalc::m_hitData`.


Former-commit-id: ad6d7a19c3bf4a5b51081d648ecc4381a761a94f