Migrate MuonDetailedTrackTruthMaker to use DataHandles
This commit migrates the MuonDetailedTrackTruthMaker algorithm to use handles.
There is no change to the code organization: indeed, the code continues to pass containers to the Trk::DetailedTrackTruthBuilder, which may have to change in the future.
The only significant change is the addition of a method to the DetailedTrackTruthCollection that allows its member m_trackCollection to be set.
Prior to migration to handles, the DetailedTrackTruthCollections could be created using the constructor that takes a TrackCollection.
However, the WriteHandle uses (I assume) the default constructor, so the addition of a method to set the TrackCollection was necessary.
Mentioning ATLASRECTS-4116 for organizational purposes.


Former-commit-id: 652ded6b08f151e161f9b5604a8111cc8e1cd657