Move TileGeoG4SD.hh header to the src directory and clean up dependencies
Only the dependencies on `TileDetDescr` and `TileSimEvent` (along
with `GaudiKernel`) need to be public now, so I've made all other
dependencies private.  In the future it would be to good to clear
up the use of the remaining public headers in this package by
classes in the `TileGeoG4Calib` package.


Former-commit-id: 99ce2af85007b62b77272d4da715c711d8ca5df5