HepMcParticleLink migration of ISF_Event package
svnpull of changes from `ISF_Event-00-12-01`

- `ISFParticle`: adding a `setBarcodeAndUpdateHepMcParticleLink(barcode)` function taking
care of synchronizing the HMPL in case of update of the barcode.
- `src/ISFTruthIncident.cxx`: replacing calls to setBarcode() by
`setBarcodeAndUpdateHepMcParticleLink()`.
- `src/ISFTruthIncident.cxx`: trying to retrieve the `GenParticle` from
the `HepMcPartileLink` if existing when "converting" the `ISFParticle`.
- `src/ISFParticle.cxx`: Cloning `HepMcParticleLink` object when using
copy constructor for `ISFParticle`.
- Added `HepMcParticleLink` object as member variable for `ISFParticle`
and its accessors.

NB Not adding `getParentHepMcParticleLink()` accessor to `ITruthIncident.h`
as it was not possible to implement this satisfactorily for
`GeantTruthIncident`.


Former-commit-id: fb036f977376e183a87d2ca5c441d8a5e4ff4ca8