VarHandleKeys: add constructor to autodeclare Properties
Adds constructors to VarHandleKeys and VarHandleKeyArrays that call
declareProperty on their owners (eg an Algorithm or AlgTool). This
enables the use of the VarHandleKey as a Property without the need
to wrap it in a Gaudi::Property. Same for CondHandleKeys.

Modifies Algorithms in AthenaExamples/AthExHive to use this syntax.

Fixes AthAlgorithm and AthAlgTool for updated contents of
Property.documentation() string (AtlasGaudi v28r2.008).

Also modifies AthenaExamples/AthExHelloWorld to use the Gaudi::Property
style of Properties, and updates class structure to get rid of
queryInterface.


Former-commit-id: d8e7a499dbbe0dc9f640b36a1b69fa9f567063a6