Add hack to avoid unphysical values for muon TOBs (ATR-16781)
Details:
The MuonTOBs for L1Topo have 2 bits to encode the pt information (see
https://twiki.cern.ch/twiki/pub/Atlas/L1CaloUpgrade/ROD_data_format_v1.0.8.xlsx)
These should be used to encode pt values of 4, 6, and 10, with the bits '11'
reserved to indicate overflow in a given octant.
However, as reported in ATR-16781, we see MuonTOBs being fed to the
L1Topo simulation with pt>10. There might be a bug in the muon
decoding that needs to be fixed. Unfortunately it might take some time
to track down the origin of this bug, and this issue is currently
affecting the simulation of several of the high-priority B-physics
algorithms.
This is a hack that I implemented following the suggestion by @olya.
It seems to work fine and to reduce the hdw-sim mismatches of all
INVM algorithms used for B-physics, without affecting other algos.

This hack should be removed once a proper bugfix is implemented.


Former-commit-id: 60efb5384209b5a3784ebaa90d72a42337716f01