Add flag to disable muon isolation
Added a flag to turn off muon isolation (both in the isolation alg side
and in the MuonTagTool which expects the isolation variables to be filled).
This is so that the isolation can temporarily be switched off as we migrate
the muon algorithms to use data handles, which will currently cause some problems
since the isolation tools are not yet migrated. Once the isolation tools have
also been migrated, the muon isolation can be switched back on.


Former-commit-id: 1cfd1d28b829b8fdb79b89176ed3fe62224862cc