Sweep of DerivationFrameworkRTT-00-01-79 into 21.2
ATLASG-1202

This introduces the DerivationFrameworkRTT package into the Git repository. The latest SVN tag used in the AtlasDerivation nightlies is swept in (00-01-79). The presence of this
package is a pre-requisite for running RTT test jobs for AthDerivation, which is itself needed before any AthDerivation caches can be built. It is marked urgent for this reason.


Former-commit-id: 1602393c77d13b9eb53a94bf023b9721fbfbf63d