Trying to get pileup reweighting working again
Need to add DerivationFrameworkDataPrep back into the git release; had
only been in svn.  Updating PATJobTransforms to match the trunk in svn
(git was quite out of date).  Hoping these changes are ok in the new
release.  Note that pileup reweighting depends on the pileup reweighting
tool, which depends on the MC event weight tool, which lives in reweight
utils, which depends on LHAPDF, which isn't available in the release at
the moment (it is an external).  Marking this a WIP until that issue is
resolved.


Former-commit-id: 6c580a0ada027f7f7937a3b04d864c834d6422f0