Tweaks to count the number of events in a PRW file
In order to solve PRODSYS-805, we need to correctly report the number of
events 'in' a PRW file.  This is done here by checking the file format
and looking for the structure of a pileup reweighting file.  If
histograms with the right keys are found, then their number of entries
determines the number of events in the file.  The way this is done here,
we have the number of events correctly reported in the JSON file, which
should be all that is needed to get prodsys to recognize how many events
are in these files.  I didn't love the 'specific' handling of the PRW
elements of this, like the directory names, but they do seem to have
been standardized, and it didn't strike me as any worse than what's
going on in HISTEntries (which I didn't want to modify for this use
because it seems like it has a very particular purpose already, even if
the name is a bit misleading because of its generality).


Former-commit-id: 18875ea4a76c07a3e3f7209a31d15c05f3c29069