Clarifying errors and fixing ATLSUSYSW-285
e/gamma tools require us to be clear about what kind of file we are
running on, as long as autoconfiguration isn't working.  Being specific.
In the meantime, also tweaking b-tagging shower type function to give a
clear error message when used on data, and to return the only shower
type we know about in R21 (PhPy8).


Former-commit-id: 9f958c7c20cfcd1669daaf7ebbfa1c1d6c924fbb